refactor(useDefindex): extract vault balance logic to hooks/use-vault-balance.ts
Step 3/4 of #66. Moves getVaultAssetBalance, getVaultBalance,
loadUserBalances, and all balance-related state/refs into a dedicated
hook. useDefindex.ts now composes useVaultBalance() alongside
useVaultMeta(); refreshBalances and the wallet-change effect remain in
useDefindex.ts since they orchestrate across both hooks plus the cache
layer.

Verified: tsc --noEmit error count unchanged (10, same as baseline).
No consumer files modified. Cross-checked with cline.
